Extracted RelayCommand to separate file
